requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Babel==2.9.1
Django==3.2.8
django-money==2.1
Pillow==8.4.0
py-moneyed==1.2
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
